+++ divide10.h
@@ -1,5 +1,5 @@
 #ifndef DIVIDE10
 #define DIVIDE10
 #include <stdint.h>
-uint64_t divide(uint64_t, unsigned);
+uint64_t divide(uint64_t, uint64_t);
 #endif
